Removed code that was freeing the PCRE program/extra after a pcre_exec error, because all relevant error codes are issues with the string being matched, _not_ the program itself.
Also resetting m_iExecutionError to 0 with each execution, because if it failed before, you wouldn't use it again and you'd call regcomp() to get a new one (with 0 error). Now you can keep using the same t_regexp even if it had a previous execution error, so m_iExecutionError needs to be reset.